Remove page based paginator, update tests
